backend/requirements.txt
supabase==2.28.0
playwright==1.58.0
httpx==0.28.1
python-dateutil==2.9.0.post0
pytest==8.4.1
tzdata==2025.2
google-auth==2.48.0
